refactor: top app bar 공통 컴포넌트 분리
