+++ app/src/main/java/nextstep/shoppingcart/ui/common/CommonTopAppBar.kt
@@ -1,0 +1,42 @@
+package nextstep.shoppingcart.ui.common
+
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
+import androidx.compose.material3.TopAppBar
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.tooling.preview.Preview
+import nextstep.shoppingcart.R
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun CommonTopAppBar(
+    title: String,
+    onBackButtonClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    TopAppBar(
+        title = { Text(text = title) },
+        navigationIcon = {
+            IconButton(onClick = onBackButtonClick) {
+                Icon(
+                    painter = painterResource(id = R.drawable.ic_arrow_back),
+                    contentDescription = "앱바 뒤로가기"
+                )
+            }
+        },
+        modifier = modifier
+    )
+}
+
+@Preview
+@Composable
+private fun ShoppingCartTopAppBarPreview() {
+    CommonTopAppBar(
+        title = "App Bar Title",
+        onBackButtonClick = {}
+    )
+}
