+++ app/src/main/java/nextstep/shoppingcart/ui/shoppingcart/ShoppingCartScreen.kt
@@ -5,18 +5,13 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.statusBarsPadding
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import nextstep.shoppingcart.R
+import nextstep.shoppingcart.ui.common.CommonTopAppBar
 import nextstep.shoppingcart.ui.theme.ShoppingCartTheme
 
 @Composable
@@ -29,31 +24,16 @@ fun ShoppingCartScreen(
             .fillMaxSize()
             .navigationBarsPadding()
             .statusBarsPadding(),
-        topBar = { ShoppingCartTopAppBar(onBackButtonClick = onBackButtonClick) }
+        topBar = {
+            CommonTopAppBar(
+                title = stringResource(R.string.shopping_cart_app_bar_title),
+                onBackButtonClick = onBackButtonClick
+            )
+        }
     ) { innerPadding ->
         Box(modifier = Modifier.padding(innerPadding)) {
         }
     }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun ShoppingCartTopAppBar(
-    onBackButtonClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    TopAppBar(
-        title = { Text(text = stringResource(R.string.shopping_cart_app_bar_title)) },
-        navigationIcon = {
-            IconButton(onClick = onBackButtonClick) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_arrow_back),
-                    contentDescription = "장바구니 뒤로가기"
-                )
-            }
-        },
-        modifier = modifier
-    )
 }
 
 @Preview
@@ -64,8 +44,3 @@ private fun ShoppingCartScreenPreview() {
     }
 }
 
-@Preview
-@Composable
-private fun ShoppingCartTopAppBarPreview() {
-    ShoppingCartTopAppBar(onBackButtonClick = {})
-}
